requirements.txt
Django==4.2.18
Pillow==10.1.0
python-dotenv==1.0.0
django-filter==23.4
whitenoise==6.6.0
mysqlclient==2.2.0
gunicorn==21.2.0
